+++ simple-schema.js
@@ -56,7 +56,7 @@ SimpleSchema = function(schema) {
 // Inherit from Match.Where
 // This allow SimpleSchema instance to be recognized as a Match.Where instance as well
 // as a SimpleSchema instance
-SimpleSchema.prototype = new Match.Where()
+SimpleSchema.prototype = new Match.Where();
 
 // If an object is an instance of Match.Where, Meteor build-in check API will look at
 // the function named `condition` and will pass it the document to validate
@@ -65,13 +65,13 @@ SimpleSchema.prototype.condition = function(doc) {
     self.validate(doc);
     if (!self.valid())
         throw new Match.Error("One or more properties do not match the schema.");
-    return true
-}
+    return true;
+};
 
 // [backwards compatibility]
 // return a function that can be use as the second parameter of the build-in check function
 SimpleSchema.prototype.match = function() {
-    console.warn('There is no more need to call the .match() method on a object to check it.');
+    console.warn('There is no more need to call the .match() method on an object to check it.');
     console.warn('see https://github.com/aldeed/meteor-simple-schema#other-methods');
     var self = this;
     return Match.Where(function(doc) {
@@ -79,7 +79,7 @@ SimpleSchema.prototype.match = function() {
         if (!self.valid()) {
             throw new Match.Error("One or more properties do not match the schema.");
         }
-        return true
+        return true;
     });
 };
 
@@ -176,6 +176,9 @@ SimpleSchema.prototype.validateOne = function(doc, keyName) {
 
     //key must pass validation check
     var def = self._schema[keyName];
+    if (!def) {
+        throw new Error("The schema contains no key named " + keyName);
+    }
     var keyValue;
     var keyLabel = def.label || keyName;